% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_d3_method}
\alias{calc_d3_method}
\title{Calculate Day-3 Method}
\usage{
calc_d3_method(d3_cands, dx)
}
\arguments{
\item{d3_cands}{A logical vector of indices associated with dx.
d3_cands(i) is TRUE if and only if dx(i) and dx(i + 2) are positive.}

\item{dx}{A numeric vector of daily snow changes.}
}
\value{
A numeric vector of the Day-3 method. See details.
}
\description{
Takes a numeric vector of daily snow changes, as well as
a logical vector of possible indices to include in the Day-3 method,
and returns the vector of the Day-3 method.
}
\section{Details}{

An observation of the Day-3 method consists of the sum of three
consecutive elements of dx, the first and last of which must be
positive, and the sum of the three observations must be positive.
The d3_cands logical vector is TRUE when the first and third
observations are positive. At the indices where d3_cands is TRUE,
this function sums the dx(i), dx(i + 1), and dx(i + 2). If that sum
is positive, it is added as a value of the Day-3 method.
The function would then falsify d3_cands(i + 1) and d3_cands(i + 2)
as an index of dx can be used in a sum only once.
}

\examples{
dx <- c(1, -3, 3, 1, 2, 3, 1, -7, 2, -1, -2, 3)
d3_cands <- c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, TRUE, FALSE, FALSE,
 FALSE)
rdailychange:::calc_d3_method(d3_cands, dx)

}
