% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY1.R
\name{d1_method}
\alias{d1_method}
\title{Calculate D1 Method for Snow Loads}
\usage{
d1_method(df, col_name = "SWE")
}
\arguments{
\item{df}{The dataframe containing snow observations.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A list of numeric vectors containing the observations
for the D1 Method (list split across 2 or more raw observations of 0).
}
\description{
This function calculates the D1 Method.
}
