% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{d2_method_from_vector}
\alias{d2_method_from_vector}
\title{d2_method_from_vector}
\usage{
d2_method_from_vector(x)
}
\arguments{
\item{x}{A sequential change of SWE values. This is a vector object.}
}
\description{
This function take raw observations, make sure there is
enough data to perform the D2 Method (if not, it will return NULL),
and then calculate the D2 Method.
}
\examples{
x <- sample(-5:9)
rdailychange:::d2_method_from_vector(x)
}
