% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day3.R
\name{d3_method}
\alias{d3_method}
\title{Calculate D3 Method for Snow Loads}
\usage{
d3_method(df, col_name = "SWE")
}
\arguments{
\item{df}{The dataframe containing snow observations.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A list of numeric vectors containing the observations
for the D3 Method (list split across 2 or more raw observations of 0).
}
\description{
Given a dataframe, a column of which includes snow observations,
this function will calculate the D3 Method.
}
\examples{
ID <- rep("USW00023062", 13)
NAME <- rep("DENVER-STAPLETON", 13)
STATE <- rep("CO", 13)
LATITUDE <- rep(39.7633, 13)
LONGITUDE <- rep(-104.8694, 13)
DATE <- c(1950-11-08, 1950-11-09, 1950-11-10, 1950-11-11, 1950-11-12,
          1950-11-13, 1950-12-05, 1950-12-06, 1950-12-07, 1950-12-08,
          1951-01-06, 1951-01-07, 1951-01-08)
SWE <- c(22.553862, 29.897561, 15.685390, 11.953282, 8.247274, 4.224420,
         13.676186, 13.737334, 9.453138, 4.829772, 20.319270, 10.564117,
      10.595503)
sample_data <- data.frame(ID, NAME, STATE, LATITUDE, LONGITUDE, DATE, SWE)
rdailychange:::d3_method(sample_data)
}
