% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day3.R
\name{d3_method_from_vector}
\alias{d3_method_from_vector}
\title{d3_method_from_vector}
\usage{
d3_method_from_vector(x)
}
\arguments{
\item{x}{A sequential change of SWE values. This is a vector object.}
}
\description{
This function takes raw observations, make sure there is
enough data to perform the D3 Method (if not, it will return NULL),
and then calculate the D3 Method.
}
