% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day4.R
\name{d4_method}
\alias{d4_method}
\title{D4 Method for Snow Loads}
\usage{
d4_method(df, col_name = "SWE")
}
\arguments{
\item{df}{The dataframe containing snow observations.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A list of numeric vectors containing the observations
for the D4 Method (list split across 2 or more raw observations of 0).
}
\description{
Given a dataframe, a column of which includes snow observations,
this function will calculate the D4 Method.
}
