% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day4.R
\name{d4_method_from_vector}
\alias{d4_method_from_vector}
\title{d4_method_from_vector}
\usage{
d4_method_from_vector(x)
}
\arguments{
\item{x}{A vector of SWE values}
}
\description{
This function will take raw observations, make sure there is
enough data to perform the D4 Method (if not, it will return NULL), and
then calculate the D4 Method.
}
