% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day5.R
\name{d5_method_from_vector}
\alias{d5_method_from_vector}
\title{d5_method_from_vector}
\usage{
d5_method_from_vector(x)
}
\arguments{
\item{x}{A sequential change of SWE values. This is a vector object.}
}
\description{
This function will take raw observations, make sure there is
enough data to perform the D5 Method (if not, it will return NULL), and then
calculate the D5 Method.
}
