% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{day2}
\alias{day2}
\title{Extract Day-2 method observations}
\usage{
day2(station_data, col_name)
}
\arguments{
\item{station_data}{A data table of a measurement location/station.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A list with two elements. The first element is the Day-2
method observations, while the second element is the annual maximum load
for each snow year.
}
\description{
This function allows the user to extract sequential daily
changes
(Day-1 method) in SWE.
}
