% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day3.R
\name{day3}
\alias{day3}
\title{Extract Day-3 method observations}
\usage{
day3(station_data, col_name)
}
\arguments{
\item{station_data}{A data table of a measurement location/station.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A list with two elements. The first element is the Day-3
method observations, while the second element is the annual maximum load
for each snow year.
}
\description{
This function allows the user to extract sequential daily
changes
(Day-1 method) in SWE.
}
\examples{
sample_data <- rdailychange::sample_data
sample_data <- sample_data[ID == "USW00023062"]
rdailychange:::day3(sample_data, col_name = "SWE")
}
