% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_observations.R
\name{extract_observations}
\alias{extract_observations}
\title{Extraction of Daily Change Observations}
\usage{
extract_observations(df, day = 1, col_name = "SWE")
}
\arguments{
\item{df}{data frame of SWE and other measurement location metadata.}

\item{day}{Daily change method. This is an integer parameter. Default is
1 for Day-1 method.}

\item{col_name}{Character string of the column name containing the
snow observations.}
}
\value{
A data table of measurement location's metadata and specified
daily sequential observations.
}
\description{
This function allows the user to extract sequential daily
changes from Snow Water Equivalent (SWE) observations. Day-1 observations
are sequential daily changes in SWE.
Day-2 observations assume that the shedding phenomenon occurs after two days.
Hence, Day-2 observations are a sum of two consecutive days of positive
sequential changes. Day-3, Day-4 and Day-5 observations are a sum of three,
four and five consecutive sequential daily changes respectively. In addition,
they allow for at most one middle sequential change to be a negative value as
long as the sum of the other sequential daily changes are greater than the
negative change.
}
\examples{
sample_data <- rdailychange::sample_data
sample_data <- sample_data[ID == "USW00023062"]
rdailychange:::extract_observations(sample_data, day = 2)

}
