% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp_gev.R
\name{fit_gp_gev}
\alias{fit_gp_gev}
\title{Helper Function to Fit Extracted Observations}
\usage{
fit_gp_gev(list, type = "GEV")
}
\arguments{
\item{list}{A list object of extracted observations.}

\item{type}{The type of distribution (GEV or GP) to fit to the observations.
The default is GEV.}
}
\value{
A fit summary with measurement location metadata , MRIs and Annual
MRI ratios
}
\description{
This function allows the user to fit a Generalized Extreme Value
(GEV) or a Generalized Pareto (GP) distribution to the
extracted observations. the function does not fit Measurement locations
with less than 30.
}
