% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_observations}
\alias{fit_observations}
\title{Fit Extracted Observations}
\usage{
fit_observations(list, type = "GEV")
}
\arguments{
\item{list}{A list object of extracted observations.}

\item{type}{The type of distribution (GEV or GP) to fit to the observations.
The default is GEV.}
}
\value{
A fit summary with measurement location metadata , MRIs and Annual
MRI ratios.
}
\description{
This function allows the user to fit a Generalized Extreme Value
(GEV) or a Generalized Pareto (GP) distribution to the
extracted observations. The function does not fit Measurement locations
with less than 30.
}
\examples{
sample_data <- rdailychange::sample_data
sample_data <- sample_data[ID == "USW00023062"]
t <- extract_observations(sample_data, day = 2)
rdailychange:::fit_observations(t)

}
