% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_events.R
\name{plot_events}
\alias{plot_events}
\title{Map of MRIs or Annual MRIs Ratio}
\usage{
plot_events(df, size = 2.5, event = 50, annual = FALSE)
}
\arguments{
\item{df}{A data frame of a fit summary object.
The fit summary includes MRIs or Annual MRIs Ratio of the respective
measurement locations.}

\item{size}{Size of points on the map. This is an integer value that
corresponds to the diameter of the points.}

\item{event}{MRI/event values to map. The default is 50, which maps the
50-year.
MRI/event values. If the user specifies the value 100 or 500, the map will
show the 100-year or 500-year MRI/event values respectively.}

\item{annual}{Annual MRI Ratios of the respective measurement locations
are mapped. the default is FALSE.}
}
\value{
USA map of MRIs or Annual MRIs Ratio.
}
\description{
This function plots the MRIs or Annual MRIs Ratio on a USA map.
}
