% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdailychange-package.R
\docType{package}
\name{rdailychange-package}
\alias{rdailychange}
\alias{rdailychange-package}
\title{rdailychange: Daily Change Method for Spatial Data}
\description{
Spatial analysis of daily snow water equivalent (SWE). You provide the spatial data in a data frame with SWE and other specified columns, tell "rdailychange" which daily change to extract, which distribution(GEV or GP) to fit to the extracted observations, and how to map the ground snow load mean recurrence intervals (MRIs) or return levels.
}
\author{
\strong{Maintainer}: Kenneth Kin Pomeyie \email{kinekenneth48@gmail.com}

Authors:
\itemize{
  \item Braden Mindrum \email{braden.mindrum@usu.edu}
  \item Kinspride Duah \email{kinspride2020@gmail.com}
}

}
\keyword{internal}
