% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY1.R
\name{split_across_n_m}
\alias{split_across_n_m}
\title{split_across_n_m}
\usage{
split_across_n_m(x, n = 2, m = 0)
}
\arguments{
\item{x}{A sequential change of SWE values. This is a vector object.}

\item{n}{This represents the zeros to split a vector. The default is 2, if
there are two consecutive zeros, the vector will be split at that point.}

\item{m}{This represents the value when encountered will serve as a
a splitting point.}
}
\description{
This function takes a vector x of raw snow observations and
split the vector across n or more m's into a list of vectors.
}
\examples{
x <- c(1, 2, 0, 0, 0, 4, 5, 0, 3, 0, 0, 2, 1, 3)
rdailychange:::split_across_n_m(x)
}
