% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY1.R
\name{d1_method_from_vector}
\alias{d1_method_from_vector}
\title{d1_method_from_vector}
\usage{
d1_method_from_vector(x)
}
\arguments{
\item{x}{A sequential change of SWE values. This is a vector object.}
}
\description{
This function works in the same manner as Helper Function 1
but this time time the vector x has no consecutive 0's. This function takes
vector x of raw observations with no consecutive 0's, make sure there is
enough data to perform the D1 Method (if not, it will return NULL), and then
calculate the D1 Method.
}
\examples{
x <- c(1, 2, 0, 4, 5, 0, 3, 0, 2, 1, 3)
rdailychange:::d1_method_from_vector(x)
}
