% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day4.R
\name{middle_negative_test}
\alias{middle_negative_test}
\title{middle_negative_test}
\usage{
middle_negative_test(dx)
}
\arguments{
\item{dx}{A sequential change of SWE values. This is a vector object.}
}
\description{
For the D4 Method, we need to have four consecutive changes,
the ends of which are positive, and at most one of the two middles is
negative. This function takes the changes and returns a logical vector of
TRUEs when at most one of the two middles is negative.
}
\examples{
x <- c(1, -2, -1, 0, 0, -4, 5, 0, 3, -1, 0, -2, -1, 3)
rdailychange:::middle_negative_test(x)
}
