% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Annual snow load maximums of Utah weather stations.}
\format{
A data frame with 16081 rows and 5 variables:
\describe{
\item{ID}{This represents the station's ID.}
\item{NAME}{Weather Station Name.}
\item{STATE}{The state of the weather station.}
\item{LATITUDE}{The angular distance of the station north or south of the
earth's equator. This is expressed in degrees.}
\item{LONGITUDE}{The angular distance of the weather east or west of the
meridian at Greenwich, England, or west of the standard meridian of a
celestial object. This is expressed in degrees.}
\item{DATE}{This is the date the snow water equivalent was recorded.}
\item{SWE}{This is the snow water equivalent value recorded.}
}
}
\source{
\url{https://www.ncei.noaa.gov/data/global-historical-climatology-network-daily/doc/GHCND_documentation.pdf}
}
\usage{
sample_data
}
\description{
This data contains 15,172 seasonal snow load maximums of
389 weather stations in Utah. The snow load values are measured in
kilopascal (kPa). For WESD measurements, the associated values represent
the direct measurement of snow load while SNWD values represent an
estimate of the snow load.
}
\keyword{datasets}
